# Back-End/Api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Phone Tracker API
  description: |
    Phone Tracker API for academic misconduct monitoring and student management.
    
    This API provides comprehensive endpoints for:
    - User authentication and session management
    - Student and system user management
    - Occurrence tracking and feedback systems
    - Analytics and reporting features
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Permissions
    The API uses role-based access control with the following roles:
    - **ADMIN**: Full system access
    - **STAFF**: Management operations
    - **STUDENT**: Limited access to own data
    - **TEACHER**: Educational tools access
    
    ## Error Handling
    The API uses standard HTTP status codes and returns detailed error information in JSON format.
    All error responses follow a consistent structure with timestamp, status, error type, message, and path.

  version: 1.0.0
  contact:
    name: Luiz Gustavo
    email: luiz.gustavo@aluno.ifsp.edu.br
    url: https://github.com/luizgustavosgobi
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    ChangePasswordDto:
      type: object
      required:
        - id
        - newPassword
      properties:
        id:
          type: string
          description: User or student ID whose password will be changed
          minLength: 3
          maxLength: 50
          example: "admin001"
        oldPassword:
          type: string
          format: password
          description: Current password (optional, used for self-service password changes)
          minLength: 6
          maxLength: 100
        newPassword:
          type: string
          format: password
          description: New password (must meet security requirements)
          minLength: 8
          maxLength: 100
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]'
          example: "NewSecure123!"
      description: Password change request data

    # Student Management DTOs
    StudentDto:
      type: object
      required:
        - id
        - name
        - course
      properties:
        id:
          type: string
          description: Unique student identifier (typically student ID number)
          minLength: 3
          maxLength: 20
          pattern: '^[a-zA-Z0-9]+$'
          example: "20240001"
        name:
          type: string
          description: Full name of the student
          minLength: 2
          maxLength: 100
          pattern: '^[a-zA-ZÀ-ÿ\s]+$'
          example: "Alice Johnson"
        course:
          type: string
          description: Student's course or program of study
          minLength: 2
          maxLength: 100
          example: "Computer Science"
        photo:
          type: string
          format: byte
          description: Base64 encoded profile photo (optional, max 2MB)
          maxLength: 2800000
          example: "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQ..."
      description: Student account information for academic management
    StudentResponseDto:
      allOf:
        - $ref: '#/components/schemas/StudentDto'
        - type: object
          properties:
            createdAt:
              type: string
              format: date-time
              description: Account creation timestamp (UTC)
              example: "2024-01-01T12:00:00.000Z"
      description: Student information returned in API responses
    StudentAndFeedbackResponseDto:
      allOf:
        - $ref: '#/components/schemas/StudentResponseDto'
        - type: object
          properties:
            occurrences:
              type: array
              description: Recent feedback/occurrences for this student (max 10 most recent)
              maxItems: 10
              items:
                $ref: '#/components/schemas/FeedbackSummaryDto'
            totalOccurrences:
              type: integer
              description: Total number of occurrences for this student
              minimum: 0
              example: 5
            approvedOccurrences:
              type: integer
              description: Number of approved occurrences
              minimum: 0
              example: 3
            rejectedOccurrences:
              type: integer
              description: Number of rejected occurrences
              minimum: 0
              example: 2
            pendingOccurrences:
              type: integer
              description: Number of occurrences awaiting review
              minimum: 0
              example: 0
            lastOccurrenceDate:
              type: string
              format: date-time
              description: Date of most recent occurrence (null if no occurrences)
              nullable: true
              example: "2024-01-15T14:30:00.000Z"
      description: Complete student profile with occurrence statistics and recent feedback
    FeedbackSummaryDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique feedback identifier
          example: "123e4567-e89b-12d3-a456-426614174000"
        comments:
          type: string
          description: Staff comments about the occurrence
          maxLength: 1000
          example: "Valid occurrence confirmed"
        status:
          type: string
          enum: [APPROVED, REJECTED, PENDING]
          description: Review status of the occurrence
          example: "APPROVED"
        reviewDate:
          type: string
          format: date-time
          description: When the review was completed (null if pending)
          nullable: true
          example: "2024-01-15T15:00:00.000Z"
      description: Summary of feedback for student profile display

    # Occurrence Management DTOs
...
    OccurrenceResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique occurrence identifier
          example: "123e4567-e89b-12d3-a456-426614174000"
        local:
          type: string
          description: Location where the occurrence happened
          example: "Classroom A"
        date:
          type: string
          format: date-time
          description: When the occurrence happened (UTC)
          example: "2024-01-15T14:30:00.000Z"
        type:
          type: string
          enum: [PHONE_USAGE, SLEEPING, TALKING, EATING, OTHER]
          description: Type of misconduct detected
          example: "PHONE_USAGE"
        cameraId:
          type: string
          description: Camera identifier that detected the occurrence
          example: "CAM_001"
        proofFile:
          type: string
          description: Filename of the proof file (null if no proof available)
          nullable: true
          example: "proof_20240115_143000.jpg"
        confidence:
          type: number
          format: float
          description: AI confidence level (0.0 to 1.0, null if not AI-detected)
          minimum: 0.0
          maximum: 1.0
          nullable: true
          example: 0.95
        feedback:
          $ref: '#/components/schemas/FeedbackResponseDto'
      description: Complete occurrence information with optional feedback
...
    FeedbackResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique feedback identifier
          example: "456e7890-f12a-34b5-c678-901234567890"
        comments:
          type: string
          description: Staff comments about the occurrence
          nullable: true
          example: "Valid occurrence confirmed"
        status:
          type: string
          enum: [APPROVED, REJECTED, PENDING]
          description: Review status
          example: "APPROVED"
        reviewedBy:
          type: string
          description: ID of staff member who provided the feedback
          nullable: true
          example: "staff001"
        reviewDate:
          type: string
          format: date-time
          description: When the feedback was submitted (UTC)
          nullable: true
          example: "2024-01-15T15:00:00.000Z"
        student:
          $ref: '#/components/schemas/StudentSummaryDto'
      description: Complete feedback information with reviewer details
    StudentSummaryDto:
      type: object
      properties:
        id:
          type: string
          description: Student unique identifier
          example: "20240001"
        name:
          type: string
          description: Student full name
          example: "Alice Johnson"
      description: Basic student information for feedback associations
...
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
          description: When the error occurred (UTC)
          example: "2024-01-01T12:00:00.000Z"
        status:
          type: integer
          description: HTTP status code
          minimum: 400
          maximum: 599
          example: 400
        error:
          type: string
          description: HTTP status reason phrase
          example: "Bad Request"
        message:
          type: string
          description: Detailed error message explaining what went wrong
          example: "Validation failed for field 'name': must not be blank"
        path:
          type: string
          description: API endpoint that generated the error
          example: "/users/create"
        details:
          type: object
          description: Additional error details (validation errors, etc.)
          nullable: true
          additionalProperties: true
          example:
            field: "name"
            rejectedValue: ""
            code: "NotBlank"
      description: Standard error response format used across all endpoints
...
paths:
...
  /students:
    post:
      tags:
        - Students
      summary: Create a new student account
      description: |
        Creates a new student account in the system with academic information.
        
        A default password is automatically generated based on the student's first name and ID 
        in the format: {firstName}_{studentId}
        
        **Required Permissions:** `ROLE_ADMIN` or `ROLE_STAFF`
        
        **Business Rules:**
        - Student ID must be unique across the system
        - Password is auto-generated and should be changed on first login
        - Photo is optional and can be added later
      operationId: createStudent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentDto'
      responses:
        '200':
          description: Student account created successfully
        '400':
          description: Invalid request body or validation errors
        '401':
          description: Authentication required
        '403':
          description: Access denied - User does not have ADMIN or STAFF role
        '409':
          description: Conflict - Student with the provided ID already exists
  /students/{id}:
    get:
      tags:
        - Students
      summary: Retrieve student information by ID
      description: |
        Retrieves detailed information for a specific student, including their profile and summary statistics.
        
        **Required Permissions:** 
        - `ROLE_ADMIN` - Can view any student's information
        - `ROLE_STAFF` - Can view any student's information
        - Students can only view their own information
        
        **Data Returned:** Complete student profile including academic information, occurrence statistics, and feedback summary.
      operationId: getStudentInfo
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the student to retrieve
          schema:
            type: string
          example: "20240001"
      responses:
        '200':
          description: Student information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentAndFeedbackResponseDto'
        '401':
          description: Authentication required
        '403':
          description: Access denied
        '404':
          description: Student not found
  /students/{id}/occurrences:
    get:
      tags:
        - Students
      summary: Retrieve student occurrences
      description: |
        Retrieves a paginated list of occurrences for a specific student, with optional filters.
        
        **Required Permissions:** 
        - `ROLE_ADMIN` - Can view any student's occurrences
        - `ROLE_STAFF` - Can view any student's occurrences
        - Students can only view their own occurrences
        
        **Features:**
        - Paginated results sorted by date (newest first)
        - Configurable filters for date range, occurrence type, etc.
        - Includes associated feedback and proof files
      operationId: getStudentOccurrences
      parameters:
        - name: id
          in: path
          required: true
          description: Student ID to retrieve occurrences for
          schema:
            type: string
          example: "20240001"
        - name: startDate
          in: query
          description: Filter occurrences from this date (YYYY-MM-DD)
          schema:
            type: string
            format: date
          example: "2024-01-01"
        - name: endDate
          in: query
          description: Filter occurrences until this date (YYYY-MM-DD)
          schema:
            type: string
            format: date
          example: "2024-12-31"
        - name: type
          in: query
          description: Filter by occurrence type
          schema:
            type: string
          example: "PHONE_USAGE"
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          schema:
            type: string
            default: "date,desc"
      responses:
        '200':
          description: Student occurrences retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/OccurrenceResponseDto'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
        '401':
          description: Authentication required
        '403':
          description: Access denied
        '404':
          description: Student not found
  /students/delete:
    delete:
      tags:
        - Students
      summary: Delete a student account
      description: |
        Permanently removes a student account from the system by their ID.
        
        **Required Permissions:** `ROLE_ADMIN` or `ROLE_STAFF`
        
        **Warning:** This action is irreversible. All student data including occurrences and feedback will be permanently deleted.
      operationId: deleteStudent
      parameters:
        - name: id
          in: query
          required: true
          description: Unique identifier of the student to be deleted
          schema:
            type: string
          example: "20240001"
      responses:
        '200':
          description: Student account deleted successfully
        '401':
          description: Authentication required
        '403':
          description: Access denied
        '404':
          description: Student not found
  /students/edit:
    put:
      tags:
        - Students
      summary: Update student information
      description: |
        Updates a student's academic information such as name, course, and photo.
        
        **Required Permissions:** `ROLE_ADMIN` or `ROLE_STAFF`
        
        **Note:** Password cannot be changed through this endpoint. Use /students/change_password instead.
      operationId: editStudent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentDto'
      responses:
        '200':
          description: Student information updated successfully
        '400':
          description: Invalid request body or validation errors
        '401':
          description: Authentication required
        '403':
          description: Access denied
        '404':
          description: Student not found
  /students/change_password:
    put:
      tags:
        - Students
      summary: Change student password
      description: |
        Updates the password for a specific student account.
        
        **Required Permissions:** `ROLE_ADMIN` or `ROLE_STAFF`
        
        **Security Notes:**
        - The new password will be securely hashed before storage
        - Student should be notified of the password change through appropriate channels
      operationId: changeStudentPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordDto'
      responses:
        '200':
          description: Student password changed successfully
        '400':
          description: Invalid request body or validation errors
        '401':
          description: Authentication required
        '403':
          description: Access denied
        '404':
          description: Student not found
  /students/autocomplete:
    get:
      tags:
        - Students
      summary: Student search autocomplete
      description: |
        Provides autocomplete functionality for searching students by ID or name.
        Returns a list of students whose ID or name starts with the provided query string.
        
        **Required Permissions:** `ROLE_ADMIN`, `ROLE_STAFF` or `ROLE_TEACHER`
        
        **Use Cases:**
        - Search bars in occurrence reporting forms
        - Quick student lookup in administrative interfaces
        - Teacher tools for classroom management
        
        **Performance:** Optimized for fast response times with limited result sets.
      operationId: studentAutoComplete
      parameters:
        - name: query
          in: query
          required: true
          description: Search query - beginning of student ID or name
          schema:
            type: string
          example: "alice"
      responses:
        '200':
          description: Autocomplete results retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentResponseDto'
        '401':
          description: Authentication required
        '403':
          description: Access denied

  # Occurrence Management endpoints
...
  /proofs/{fileName}:
    get:
      tags:
        - Proofs
      summary: Retrieve proof file by filename
      description: |
        Retrieves a proof file (evidence) by its filename for authenticated users.
        
        **Required Permissions:** None - Public endpoint (Has to be authenticated)
        
        **Use Cases:**
        - Staff reviewing occurrence evidence
        - Administrative access to proof files
        - Secure file serving with authentication
        
        **Security Features:**
        - Filename validation and sanitization
        - Path traversal protection
        - Content type detection and proper headers
        
        **File Types Supported:** Images, videos, documents, and other multimedia evidence files.
      operationId: getProof
      parameters:
        - name: fileName
          in: path
          required: true
          description: Name of the proof file to retrieve
          schema:
            type: string
            pattern: '^[a-zA-Z0-9._-]+$'
            maxLength: 255
          example: "proof_20240115_143000.jpg"
      responses:
        '200':
          description: Proof file retrieved successfully
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
                description: Binary content of the proof file
            image/*:
              schema:
                type: string
                format: binary
                description: Image proof file
            video/*:
              schema:
                type: string
                format: binary
                description: Video proof file
            application/pdf:
              schema:
                type: string
                format: binary
                description: PDF document proof file
          headers:
            Content-Type:
              description: MIME type of the file
              schema:
                type: string
                example: "image/jpeg"
            Content-Length:
              description: Size of the file in bytes
              schema:
                type: integer
                example: 2048576
        '401':
          description: Authentication required - Missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied - User does not have ADMIN or STAFF role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Proof file not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                fileNotFound:
                  summary: File not found error
                  value:
                    timestamp: "2024-01-01T12:00:00.000Z"
                    status: 404
                    error: "Not Found"
                    message: "File not found"
                    path: "/proofs/proof_20240115_143000.jpg"
        '500':
          description: Server error - Failed to read or serve the file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...